feat: Add initialization of global force vector
This commit adds the initialization of the global force vector in the `assembleBoundaryFem.m` file. The global force vector is now initialized as a sparse matrix with dimensions equal to the number of nodes multiplied by the dimensionality of the problem.

feat: Add controller loads

This commit adds support for controller loads in the `assembleBoundaryFem.m` file. If a controller function is defined in the system, it calculates the control input and multiplies it with an input map to obtain the controller forces. If no controller function is defined but there are predefined control forces, they are multiplied with an input map to obtain the controller forces.

refactor: Remove unused code

This commit removes unused code related to assembling contact mesh in `assembleContactMeshFem.m`. The code that was commented out and not used has been removed.

fix: Fix formatting issue

This commit fixes a formatting issue in `assembleGlobalFem.m`. The spacing between two elements of a material constant array has been corrected.

refactor: Remove unnecessary comment

This commit removes an unnecessary comment from `assembleInternalForcesFem.m`. The commented line was not needed and has been removed.